Changed doubles to floats (in RealPoint and ContinuousSpace) Added default values to getNeighbors function Added function to obtaining cell of discrete grid. Added some code for further development (continuous space is not working yet)

+++ include/space/Point.hpp
@@ -11,8 +11,8 @@ struct Point {
 };
 
 struct RealPoint {
-    double x{};
-    double y{};
+    float x{};
+    float y{};
 
     bool operator==(const RealPoint& other) const {
         return x == other.x && y == other.y;
